pyproject.toml
[tool.poetry]
name = "visualizzazionecontrolli"
version = "0.1.0"
description = "Codice per la visualizzazione dei controlli formali"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
streamlit = "^1.36.0"
streamlit-aggrid = "^1.0.5"
pyarrow = "^17.0.0"
openpyxl = "^3.1.5"
streamlit-authenticator = "^0.3.3"
fitz = "^0.0.1.dev2"
st-files-connection = "^0.1.0"
boto3 = ">=1.26.0,<1.35.0"
botocore = ">=1.26.0,<1.35.0"
aiobotocore = ">=2.5.4,<3.0.0"
s3fs = ">=2023.6.1,<2025.0.0"
pdfplumber = "^0.11.2"
watchdog = "^4.0.1"
flask = "^3.0.3"
flask-cors = "^4.0.1"
python-dotenv = "^1.0.1"
pymongo = "^4.8.0"
pyopenssl = "^24.2.1"
asn1crypto = "^1.5.1"
pypdf2 = "^3.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
